fix: Split marketing layout into server/client components for prerender compatibility
Framer-motion useEffect fails during Next.js 16 static prerendering with React 19.
Moving dynamic='force-dynamic' to a server layout and extracting nav into a client
component prevents prerender attempts on animated marketing pages.
