+++ apps/web/src/app/(marketing)/layout.tsx
@@ -1,56 +1,11 @@
-'use client';
+import MarketingNav from './MarketingNav';
 
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-import { Button } from '@/components/ui/button';
-
-const navLinks = [
-  { href: '/landing', label: 'Home' },
-  { href: '/about', label: 'About' },
-  { href: '/roadmap', label: 'Roadmap' },
-  { href: '/docs', label: 'Docs' },
-];
+export const dynamic = 'force-dynamic';
 
 export default function MarketingLayout({ children }: { children: React.ReactNode }) {
-  const pathname = usePathname();
-
   return (
     <div className="min-h-screen bg-background">
-      <nav className="fixed top-0 left-0 right-0 z-50 border-b border-border/50 bg-background/80 backdrop-blur-xl">
-        <div className="max-w-6xl mx-auto px-6 h-14 flex items-center justify-between">
-          <Link href="/landing" className="text-lg font-bold tracking-tight">
-            <span className="bg-gradient-to-r from-primary to-[#6366F1] bg-clip-text text-transparent">
-              Siza
-            </span>
-          </Link>
-
-          <div className="hidden sm:flex items-center gap-1">
-            {navLinks.map((link) => (
-              <Link
-                key={link.href}
-                href={link.href}
-                className={`px-3 py-1.5 text-sm rounded-md transition-colors ${
-                  pathname === link.href
-                    ? 'text-foreground bg-accent'
-                    : 'text-muted-foreground hover:text-foreground'
-                }`}
-              >
-                {link.label}
-              </Link>
-            ))}
-          </div>
-
-          <div className="flex items-center gap-3">
-            <Button variant="ghost" size="sm" asChild>
-              <Link href="/signin">Sign In</Link>
-            </Button>
-            <Button size="sm" asChild>
-              <Link href="/signup">Get Started</Link>
-            </Button>
-          </div>
-        </div>
-      </nav>
-
+      <MarketingNav />
       <main className="pt-14">{children}</main>
     </div>
   );
